feat: 통계 업데이트와 대시보드 업데이트 로직 분리(#190)
