+++ src/main/java/com/dnd/namuiwiki/config/EventConfiguration.java
@@ -1,5 +1,6 @@
 package com.dnd.namuiwiki.config;
 
+import com.dnd.namuiwiki.domain.dashboard.DashboardService;
 import com.dnd.namuiwiki.domain.statistic.StatisticsService;
 import com.dnd.namuiwiki.domain.survey.SurveyEventHandler;
 import lombok.RequiredArgsConstructor;
@@ -10,10 +11,11 @@ import org.springframework.context.annotation.Configuration;
 @RequiredArgsConstructor
 public class EventConfiguration {
     private final StatisticsService statisticsService;
+    private final DashboardService dashboardService;
 
     @Bean
     public SurveyEventHandler surveyEventHandler() {
-        return new SurveyEventHandler(statisticsService);
+        return new SurveyEventHandler(statisticsService, dashboardService);
     }
 
 }
